% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_size.R
\name{adjusted_size}
\alias{adjusted_size}
\title{Corrección del tamaño de la muestra por el factor de población finita}
\usage{
adjusted_size(samp_size, pob_size)
}
\arguments{
\item{samp_size}{valor del tamaño de la muestra inicial}

\item{pob_size}{valor del tamaño de la población}
}
\value{
n0 tamaño de muestra ajustado
}
\description{
Función que corrige el tamaño de la muestra cuando n/N>0.05
}
\examples{
 adjusted_size(385,500)
}
