% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ausentismo}
\alias{ausentismo}
\title{Ausencias al trabajo}
\format{
Data con 48 registros y 7 variables
\describe{
\item{id}{identificador del trabajador}
\item{ausen}{dias ausentes del trabajo por año}
\item{taller}{1 trabaja en el taller, 0 no trabaja en el taller}
\item{sexo}{1 hombre, 0 mujer}
\item{edad}{edad del trabajador en años}
\item{antg}{antiguedad del trabajador en años}
\item{sala}{salario del trabajador en miles de pesos}
}
}
\usage{
ausentismo
}
\description{
Esta data contiene caracteristicas de trabajadores que faltan a su trabajo
}
\examples{
ausentismo
}
\keyword{datasets}
