% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casas}
\alias{casas}
\title{Pricios y caracteristicas de viviendas.}
\format{
Una data con 54 filas y 10 variables:
\describe{
\item{id}{consecutivo}
\item{precio}{precio de la vivienda en miles de U$D}
\item{imp.pred}{valor impuesto predial}
\item{num.ban}{número de baños}
\item{area.lot}{área del lote}
\item{area.con}{área construida}
\item{num.gar}{número de garajes}
\item{num.cua}{número de cuartos}
\item{num.hab}{número de habitaciones}
\item{edad}{edad de la vivienda}
...
}
}
\usage{
casas
}
\description{
Esta data contiene los precios y principales caraceristica de vienes inmuebles
}
\examples{
casas
}
\keyword{datasets}
