% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cnk.R
\name{Cnk}
\alias{Cnk}
\title{Combinación}
\usage{
Cnk(n, k)
}
\arguments{
\item{n}{tamaño de la población, número entero}

\item{k}{tamaño de la muestra, número entero}
}
\value{
Cnk  combinación de n con k
}
\description{
Función que calcula n permutado k
}
\examples{
Cnk(10,2)

}
