% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalo.var.R
\name{intervalo.var}
\alias{intervalo.var}
\title{Intervalo de confianza para una varianza}
\usage{
intervalo.var(x, niv.conf = 0.95)
}
\arguments{
\item{x}{vector con variable numérica}

\item{niv.conf}{nivel de confianza . p.ej:95}
}
\value{
ic.var intervalo de confiaza para una varianza
}
\description{
Función que calcula Calcula intervalo de confianza para una varianza :
}
\examples{
intervalo.var(c(3,4,3,2,3,4,5),0.95)
}
