% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paqueteDEG-package.R
\docType{package}
\name{paqueteDEG-package}
\alias{paqueteDEG}
\alias{paqueteDEG-package}
\title{paqueteDEG: Paquete curso Probabilidad y Estadistica}
\description{
Paquete creado como apoyo al curso de probabilidad y estadística, Pontificia Universidad Javeriana Cali. Contiene tambien tutoriales como apoyo a las monitorias del curso para los diferentes módulos.
}
\author{
\strong{Maintainer}: Daniel Enrique Gonzalez Gomez \email{dgonzalez@javerianacali.edu.co} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
