% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pnk.R
\name{Pnk}
\alias{Pnk}
\title{Permutación}
\usage{
Pnk(n, k)
}
\arguments{
\item{n}{tamaño de la población, número entero}

\item{k}{tamaño de la muestra, número entero}
}
\value{
Pnk  permutacion de n con k
}
\description{
Función que calcula n permutado k
}
\examples{
Pnk(10,2)

}
